Consolidate duplicate test suites onto one file per subject
Three subjects were each covered by two test files. Test counts are preserved
in every case; what goes away is the redundant half.

add_prompt
    test_add_prompt_to_marketplace.py folded into
    test_agent_marketplace_handler.py. Eleven of its thirteen tests were
    already covered there, and covered better: it mocked httpx.Client
    directly, while the surviving file routes through httpx.MockTransport so
    real request construction, URL encoding and status handling stay in the
    loop. The two genuinely uncovered cases were ported rather than copied:
    timeout propagation on POST (mirroring the existing GET test, with
    install_post_transport now recording timeouts the way
    install_get_transport already did), and 401, by parametrising the 4xx test
    over 400/401/403/404. 69 tests -> 77.

litellm
    tests/utils/test_litellm_wrapper.py is now the single home, holding what
    was tests/test_multi_provider.py. The 444 lines it replaces collected
    exactly one test, and that test asserted nothing: every check sat inside a
    try/except that printed a mark and continued, so it could not fail. Its
    module docstring now records what that script walked through and what is
    therefore still unasserted -- streaming via arun, batched_run, tool
    calling, vision detection, direct-URL images. 1 real test -> 31.

one-on-one debate
    New tests/structs/test_one_on_one_debate.py covers both implementations of
    the same feature: deep_discussion.one_on_one_debate (function) and
    multi_agent_debates.OneOnOneDebate (class), sectioned by API so they
    cannot drift apart unnoticed. Two test names collided across the old
    files; the class-API copies carry a _class_ infix.
    test_multi_agent_debate.py keeps its five ExpertPanelDiscussion tests.
    18 + 11 -> 24 + 5.

Also fixes an import-time crash in initialize_logger. The log folder defaulted
to os.getenv("WORKSPACE_DIR"), which a default argument evaluates once at
import; with WORKSPACE_DIR unset that is None, and os.path.exists(None) raises
TypeError, so importing swarms failed for anyone without the variable set. The
value is now resolved per call with a "logs" fallback, keeping the
WORKSPACE_DIR behaviour when it is set.

Carried along, not authored here -- flagged so a reviewer can scope them:

  - hybrid_hiearchical_peer_swarm.py drops its local convert_str_to_dict in
    favour of the shared swarms.utils.str_to_dict helper
  - func_to_str.py builds its string with a list and join
  - get_reasoning_efforts.py shortens its comments and docstring
  - tests/benchmarks/aop_benchmark.py deleted (2,935 lines). This was the last
    thing exercising swarms/structs/aop.py, which now has no coverage of any
    kind
  - tests/utils/test_str_to_dict.py deleted (168 lines), in the same change
    that gives str_to_dict a second caller
